Fix compile (#80)
* Fix compile

* Fix compile

* Fix mtx in fb effects

---------

+++ courses/koopa_troopa_beach/course_offsets.c
@@ -316,7 +316,7 @@ Animation* d_course_koopa_troopa_beach_unk_data5[] = {
 };
 
 // No idea what this stuff is
-u32 d_course_koopa_troopa_beach_unk_data6[] = {
+uintptr_t d_course_koopa_troopa_beach_unk_data6[] = {
     0x00000000, 0x00000000, 0x00000001, 0x00000000, 0x00000001,
 };
